requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
loguru>=0.7.0
pydantic>=2.0.0